+++ packages/shared/selectors.ts
@@ -1,8 +1,5 @@
 import { Scene } from '@dcl/schemas'
-import { parseParcelPosition } from 'atomicHelpers/parcelScenePositions'
-import { ETHEREUM_NETWORK, getAssetBundlesBaseUrl } from 'config'
-import { ContentMapping, LoadableParcelScene, SceneFeatureToggle } from './types'
-import { SceneWorker } from './world/SceneWorker'
+import { ContentMapping, SceneFeatureToggle } from './types'
 
 export function normalizeContentMappings(
   mappings: Record<string, string> | Array<ContentMapping>
@@ -20,27 +17,6 @@ export function normalizeContentMappings(
   }
 
   return ret
-}
-
-/**
- * This is the format of scenes that needs to be sent to Unity to create its counterpart
- * of a SceneWorker
- */
-export function sceneWorkerToLoadableParcelScene(worker: SceneWorker): LoadableParcelScene {
-  const entity = worker.loadableScene.entity
-  const mappings: ContentMapping[] = normalizeContentMappings(entity.content)
-
-  return {
-    id: worker.loadableScene.id,
-    sceneNumber: worker.rpcContext.sceneData.sceneNumber,
-    basePosition: parseParcelPosition(entity.metadata?.scene?.base || '0,0'),
-    name: getSceneNameFromJsonData(entity.metadata),
-    parcels: entity.metadata?.scene?.parcels?.map(parseParcelPosition) || [],
-    baseUrl: worker.loadableScene.baseUrl,
-    baseUrlBundles: getAssetBundlesBaseUrl(ETHEREUM_NETWORK.MAINNET) + '/',
-    contents: mappings,
-    loadableScene: worker.loadableScene
-  }
 }
 
 export function getOwnerNameFromJsonData(jsonData?: Scene) {
